KVM: x86/mmu: Init memslot hugepage information if memory attributes in use
The list of supported memory attributes is about to grow, and they will
be available on any VM type (as opposed to only ones targeted at
confidential computing). As such, update the check in
kvm_mmu_init_memslot_memory_attributes() to initialize huge page
information if any kind of memory attributes is in use.

Signed-off-by: Nicolas Saenz Julienne <nsaenz@amazon.com>