KVM: x86: Implement Hyper-V suspended state
Implement a suspension state for Hyper-V enlightened vCPUs. The suspension
can only happen on guest instruction boundaries, because it can only start
after a VM exit. This behaviour and the suspended state itself are
specified in Microsoft's "Hypervisor Top Level Functional Specification"
(TLFS).

The suspension is implemented using a single boolean field 'suspended'
which is checked within kvm_vcpu_running() and kvm_vcpu_has_events(). This
causes the vCPU to be preempted with vcpu_block().

When the 'suspended' field in struct kvm_vcpu_hv is set, the 'waiting_on'
needs to be in a valid state and thus should not be edited. This ensures
that any CPU checking if it should wake up a vCPU, will see a valid
'waiting_on' field, making suspension thread safe. On x86 the suspension
helpers always write valid state, because write instructions with
WRITE_ONCE() are fully ordered, i.e. can't be reordered by the compiler or
the CPU.

When a vCPU is unsuspended, its 'suspended' and 'waiting_on' field are
simply cleared and a request is sent to wake up the thread it is running
on. Before it starts re-executing from where it was suspended, the vCPU
will receive any interrupts it received duruing suspension.

Signed-off-by: Nikolas Wipper <nikwip@amazon.de>