KVM: Introduce helper to query whether gfn has memory protections
Some features, like the gfn_to_pfn_cache, are incompatible with access
restriction memory attributes. Introduce
kvm_gfn_has_memory_protections() to query whether those are in use for a
specific gfn.

Memory attributes can change over time, so the caller is responsible to
synchronize with the attributes generation number if it intends to cache
this data.

Signed-off-by: Nicolas Saenz Julienne <nsaenz@amazon.com>